+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+##  2022-02-24
+
+### Added
+
+- Added Subscription Client
+
 ## [0.8.0] - 2022-01-13
 
 ### Added